Fix check_mysql_stats.py so it actually works.
I don't know how the previous commit even made it into master.  It makes the script useless. YOu could put in any host name and it'll still print STATUS OK. It never even makes a MySQL call!!

and yes this is in the 3.0.0 AND 3.0.4 version of the official pack as included in enterprise.

/facepalm