Apply changes as proposed in ZEN-16700.
In case someone does bind the template to the device or else, make sure
the exception does not occur. This commit partially applies changes
proposed in ZEN-16700, but keeps the logic of senting connection error
events only for MySQLServer component, and not for all databases.

The scope of ZEN-16700/ZEN-16585.